Revert "Unit tests are much more thorough."
